+++ src/Common/Service/ActionParamInterface.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace Soluble\MediaTools\Common\Service;
 
-use Soluble\MediaTools\Video\Exception\UnsetParamReaderException;
+use Soluble\MediaTools\Video\Exception\UnsetParamException;
 
 interface ActionParamInterface
 {
@@ -21,11 +21,16 @@ interface ActionParamInterface
     public function isParamValid(string $paramName): bool;
 
     /**
+     * Return a param, throw an exception if the param has not been defined yet or
+     * use $default if it was set.
+     *
+     * @param mixed $default Will return default value instead of throwing exception
+     *
      * @return mixed
      *
-     * @throws UnsetParamReaderException
+     * @throws UnsetParamException
      */
-    public function getParam(string $paramName);
+    public function getParam(string $paramName, $default = null);
 
     public function hasParam(string $paramName): bool;
 }